crypto-lstm-trader/requirements.txt
python-binance
pandas
numpy
ta
python-dotenv
matplotlib